InService upgrade bug fixes:
1) Fixed consecutive upgrade
2) Fixed the bug in upgrade-using-batchsize logic
3) Hold the lock while doing inService upgrade for the batch
so service.reconcile triggered by instance removal, wouldn't interfere
