+++ code/iaas/service-discovery/server/src/main/java/io/cattle/platform/servicediscovery/process/ServiceUpgrade.java
@@ -1,9 +1,6 @@
 package io.cattle.platform.servicediscovery.process;
 
 import io.cattle.platform.async.utils.TimeoutException;
-import io.cattle.platform.core.constants.CommonStatesConstants;
-import io.cattle.platform.core.constants.InstanceConstants;
-import io.cattle.platform.core.model.Instance;
 import io.cattle.platform.core.model.Service;
 import io.cattle.platform.engine.handler.HandlerResult;
 import io.cattle.platform.engine.process.ExitReason;
@@ -12,21 +9,12 @@ import io.cattle.platform.engine.process.ProcessState;
 import io.cattle.platform.engine.process.impl.ProcessExecutionExitException;
 import io.cattle.platform.json.JsonMapper;
 import io.cattle.platform.object.resource.ResourceMonitor;
-import io.cattle.platform.object.resource.ResourcePredicate;
 import io.cattle.platform.object.util.DataAccessor;
 import io.cattle.platform.process.base.AbstractDefaultProcessHandler;
-import io.cattle.platform.process.common.util.ProcessUtils;
 import io.cattle.platform.servicediscovery.api.constants.ServiceDiscoveryConstants;
 import io.cattle.platform.servicediscovery.api.dao.ServiceExposeMapDao;
-import io.cattle.platform.servicediscovery.api.util.ServiceDiscoveryUtil;
 import io.cattle.platform.servicediscovery.deployment.DeploymentManager;
 import io.cattle.platform.servicediscovery.service.ServiceDiscoveryService;
-
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.Iterator;
-import java.util.List;
-import java.util.Map;
 
 import javax.inject.Inject;
 
@@ -87,7 +75,7 @@ public class ServiceUpgrade extends AbstractDefaultProcessHandler {
     public boolean doUpgrade(Service service, io.cattle.platform.core.addon.ServiceUpgrade upgrade,
             boolean isInServiceUpgrade) {
         if (isInServiceUpgrade) {
-            return doInServiceUpgrade(service, upgrade);
+            return deploymentManager.doInServiceUpgrade(service, upgrade);
         } else {
             return doToServiceUpgrade(service, upgrade);
         }
@@ -131,80 +119,6 @@ public class ServiceUpgrade extends AbstractDefaultProcessHandler {
         }
 
         return service;
-    }
-
-    protected boolean doInServiceUpgrade(Service service, io.cattle.platform.core.addon.ServiceUpgrade upgrade) {
-        try {
-            deploymentManager.activate(service);
-
-            service = objectManager.reload(service);
-
-            long batchSize = upgrade.getBatchSize();
-
-            Map<String, List<Instance>> deploymentUnitInstancesToRemove = formDeploymentUnitsToRemove(service);
-
-            // remove deployment units
-            upgradeDeploymentUnits(batchSize, deploymentUnitInstancesToRemove, service);
-
-            if (deploymentUnitInstancesToRemove.isEmpty()) {
-                return true;
-            }
-            return false;
-
-        } catch (TimeoutException e) {
-            return false;
-        }
-    }
-
-    protected void upgradeDeploymentUnits(long batchSize, Map<String, List<Instance>> deploymentUnitInstancesToRemove, Service service) {
-        // Removal is done on per deployment unit basis
-        Iterator<Map.Entry<String, List<Instance>>> it = deploymentUnitInstancesToRemove.entrySet()
-                .iterator();
-        long i = 0;
-        while (it.hasNext() && i < batchSize) {
-            List<Instance> waitList = new ArrayList<Instance>();
-            Map.Entry<String, List<Instance>> instances = it.next();
-            for (Instance instance : instances.getValue()) {
-                objectProcessManager.scheduleProcessInstanceAsync(InstanceConstants.PROCESS_STOP,
-                        instance, ProcessUtils.chainInData(new HashMap<String, Object>(),
-                                InstanceConstants.PROCESS_STOP, InstanceConstants.PROCESS_REMOVE));
-                waitList.add(instance);
-            }
-            it.remove();
-            for (Instance instance : waitList) {
-                resourceMonitor.waitFor(instance,
-                        new ResourcePredicate<Instance>() {
-                            @Override
-                            public boolean evaluate(Instance obj) {
-                                return CommonStatesConstants.REMOVED.equals(obj.getState());
-                            }
-                        });
-            }
-            // wait for reconcile
-            deploymentManager.activate(service);
-        }
-    }
-
-    protected Map<String, List<Instance>> formDeploymentUnitsToRemove(Service service) {
-        List<String> launchConfigNames = ServiceDiscoveryUtil.getServiceLaunchConfigNames(service);
-        Map<String, List<Instance>> deploymentUnitInstancesToRemove = new HashMap<>();
-        for (String launchConfigName : launchConfigNames) {
-            List<? extends Instance> instances = exposeMapDao.listServiceManagedInstances(service, launchConfigName);
-            for (Instance instance : instances) {
-                if (!instance.getVersion().equals(
-                        ServiceDiscoveryUtil.getLaunchConfigObject(service,
-                                launchConfigName,
-                                ServiceDiscoveryConstants.FIELD_VERSION))) {
-                    List<Instance> toRemove = deploymentUnitInstancesToRemove.get(instance.getDeploymentUnitUuid());
-                    if (toRemove == null) {
-                        toRemove = new ArrayList<Instance>();
-                    }
-                    toRemove.add(instance);
-                    deploymentUnitInstancesToRemove.put(instance.getDeploymentUnitUuid(), toRemove);
-                }
-            }
-        }
-        return deploymentUnitInstancesToRemove;
     }
 
     /**
